backend/requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.0
pydantic-settings==2.2.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.36  # Updated for Python 3.13 compatibility
alembic==1.14.0
# asyncpg==0.29.0  # Temporarily disabled due to Python 3.13 compatibility
# psycopg2-binary==2.9.9  # Required for PostgreSQL connection - Commented out for SQLite dev mode

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Docker integration
docker==6.1.3

# OpenVPN integration (using subprocess for OpenVPN management)
# pyovpn==0.1.0  # This package doesn't exist, using subprocess instead

# Email
fastapi-mail==1.4.1
jinja2==3.1.2

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic-extra-types==2.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0

# Background tasks
celery==5.3.4
redis==5.0.1

# File handling
aiofiles==23.2.1

# Validation
email-validator==2.1.0
phonenumbers==8.13.26
